Well, I'm dumb
So, having a bool to handle turning on/off listening to memes doesn't quite work in this situation, as I'm trying to write this bot with more than one server in mind.
Say we have two servers, A and B, both running this bot. Server A wants memes, but server B does not. Server A could have turned it on with "ssb meme on", but because it only uses one universal bool, this would also turn on memes for server B, and subsequently, every other server running the bot. 
So ultimately, I have decided to remove this feature until I can think of a better way to implement this.